improve update_snarkos script

+++ update_snarkos.sh
@@ -5,14 +5,18 @@ upd_folder=~/aleo_update_scripts
 
 if [ -d $snarkos_folder ]; then
 	cd $snarkos_folder
-	git pull origin testnet2
-	cargo install --path .
-	cp target/release/snarkos $upd_folder/binary/snarkos
+        STATUS=$(git pull origin testnet2)
+	if [ "$STATUS" != "Already up to date." ]; then
+		cargo install --path .
+		cp target/release/snarkos $upd_folder/binary/snarkos
 	
-	cd $upd_folder
-	git add binary/
-	git commit -m "update snarkos binary"
-	git push
+		cd $upd_folder
+		git add binary/
+		git commit -m "update snarkos binary"
+		git push
+	else
+		echo "Up to date. No need rebuild."
+	fi
 else
 	echo "Can't find snarkOS flolder"
 fi
